server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2
prometheus-fastapi-instrumentator==6.1.0
websockets==12.0
pyyaml==6.0.1
